Add operands and reconstruct methods
Add operands and reconstruct methods to all AST objects.  The operands
method returns *args, **kwargs suitable to pass to the constructor.  The
reconstruct methods take *args, **kwargs and return a new object
reconstructed with the provided operands.  The former is useful in
writing generic tree visitors, the latter in writing tree-transforming
visitors.