Modify visitor infrastructure to allow passing state
Rather than the method signature indicating if the visit method is pre-
or post-order, the handler is now always responsible for visiting
children.  Handlers can now accept arbitrary *args and **kwargs which
may be used to pass state up and down the visit tree.

This may be used, for example, to pass a container for the return
values.